+++ src/main/java/com/example/mainservice/repository/UserRepository.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 
     boolean existsByEmail(String email);
 }
